move tools directory
