+++ .github/cloud-samples-tools/go.mod
@@ -1,0 +1,3 @@
+module cloud-samples-tools
+
+go 1.22.0
