+++ .github/config/nodejs.jsonc
@@ -25,6 +25,7 @@
     ".github/PULL_REQUEST_TEMPLATE.md",
     ".github/auto-label.yaml",
     ".github/blunderbuss.yaml",
+    ".github/cloud-samples-tools/",
     ".github/flakybot.yaml",
     ".github/header-checker-lint.yaml",
     ".github/snippet-bot.yml",
