--- .github/workflows/samples-tools/go.mod
@@ -1,3 +1,0 @@
-module samples-tools
-
-go 1.22.0
